+++ portfolio/app/layout.tsx
@@ -13,8 +13,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Justin Klein | Portfolio",
+  description: "Software Developer Portfolio of Justin Klein",
 };
 
 export default function RootLayout({
