+++ portfolio/components/LiveProjects.tsx
@@ -9,7 +9,7 @@ export function LiveProjects() {
             <a href="https://get-smart.justinklein.ca" target="_blank" rel="noopener noreferrer">
                 <Button 
                     variant="outlined"
-                    className="flex items-center gap-2 border border-white text-white rounded-md px-2 py-2 hover:bg-white/10 hover:translate-x-1 transition-all duration-300 focus:ring-0 normal-case text-sm font-medium tracking-wide"
+                    className="flex items-center gap-2 border border-white text-white rounded-md px-2 py-2 hover:bg-white/10 hover:translate-x-1 transition-all duration-300 focus:ring-0 normal-case md:text-md text-sm font-medium tracking-wide"
                 >
                     <span>Get Smart!</span>
                     <FaExternalLinkAlt className="h-4 w-4" />
@@ -18,18 +18,9 @@ export function LiveProjects() {
             <a href="https://get-quizzed.justinklein.ca" target="_blank" rel="noopener noreferrer">
                 <Button 
                     variant="outlined"
-                    className="flex items-center gap-2 border border-white text-white rounded-md px-2 py-2 hover:bg-white/10 hover:translate-x-1 transition-all duration-300 focus:ring-0 normal-case text-sm font-medium tracking-wide"
+                    className="flex items-center gap-2 border border-white text-white rounded-md px-2 py-2 hover:bg-white/10 hover:translate-x-1 transition-all duration-300 focus:ring-0 normal-case md:text-md text-sm font-medium tracking-wide"
                 >
                     <span>Get Quizzed!</span>
-                    <FaExternalLinkAlt className="h-4 w-4" />
-                </Button>
-            </a>
-            <a href="https://farm-games.justinklein.ca" target="_blank" rel="noopener noreferrer">
-                <Button 
-                    variant="outlined"
-                    className="flex items-center gap-2 border border-white text-white rounded-md px-2 py-2 hover:bg-white/10 hover:translate-x-1 transition-all duration-300 focus:ring-0 normal-case text-sm font-medium tracking-wide"
-                >
-                    <span>Farm Games</span>
                     <FaExternalLinkAlt className="h-4 w-4" />
                 </Button>
             </a>
