+++ portfolio/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./tech-stack"
 export * from "./project"
+export * from "./experience"
